requirements.txt
google-genai>=1.0.0
qdrant-client>=1.9.0
python-dotenv>=1.0.1
